+++ src/schedule/refresh_hot_manga_genre.js
@@ -3,11 +3,11 @@ const Hot = require("../api/hot");
 const Manga = require("../api/manga");
 
 module.exports = {
-    name: "refresh latest hot manga",
-    schedule: "0 * * * *",
+    name: "refresh hot manga genres",
+    schedule: "*/4 * * *",
     task: function () {
-        Debug.log("Refreshing latest hot mangas");
-        Hot.refreshLatest(function (ids) {
+        Debug.log("Refreshing hot mangas genres");
+        Hot.refreshAllGenres(function (ids) {
             Manga.updateAll(ids, function () {
                 Debug.log("Successfully fetched " + ids.length + " mangas");
             }, function (err) {